Refactor encoding and decoding
Lays groundwork for Arrow integration
